refactoring UI
