+++ ext-src/PolicyViolation.ts
@@ -13,17 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { PackageType } from "./PackageType";
-import { ComponentEntry } from "../ComponentEntry";
-import { RequestService } from "../RequestService";
 
-export interface PackageDependencies {
-  Dependencies: Array<PackageType>;
-  CoordinatesToComponents: Map<string, ComponentEntry>;
-  RequestService: RequestService;
-  CheckIfValid(): boolean;
-  ConvertToComponentEntry(resultEntry: any): string;
-  convertToNexusFormat(): any;
-  toComponentEntries(data: any): Array<ComponentEntry>;
-  packageForIq(): Promise<undefined>;
+ import { ConstraintViolation } from "./ConstraintViolation";
+
+export class PolicyViolation {
+  constructor(
+    readonly policyId: string,
+    readonly policyName: string,
+    readonly threatLevel: number,
+    readonly constraintViolations: Array<ConstraintViolation>
+  ) {}
 }
